s2e.c: fixed warnings

+++ linux-4.9.3/kernel/s2e/s2e.c
@@ -30,9 +30,9 @@
 
 static int __init s2e_init(void)
 {
-	size_t current_task_addr = &s2e_current_task;
-	size_t task_struct_pid_offset = offsetof(struct task_struct, pid);
-	size_t task_struct_tgid_offset = offsetof(struct task_struct, tgid);
+	uint64_t current_task_addr = (uintptr_t)&s2e_current_task;
+	uint64_t task_struct_pid_offset = offsetof(struct task_struct, pid);
+	uint64_t task_struct_tgid_offset = offsetof(struct task_struct, tgid);
 
 	if (num_online_cpus() > 1) {
 		s2e_printf("LinuxMonitor only supports single-CPU systems\n");
